Fix multi histogram weights
When using `set_average_edges` with a histogram that would have trailing zeros for weights, we weren't calculating the maximum bin position with the absolute max value, but instead the max value of that histogram's array. 

This PR fixes that issue by storing this value in a `@max` variable and updating that value when updating the edges.