Plot!
Derailed benchmarks targets older ruby versions and was using unicode_plot.rb for histograms, however that gem has a hard dependency on enumerable-statistics which is targeting Ruby 2.4+ only. This PR essentially copies all the required parts of unicode statistics for plotting a histogram without needing enumerable-statistics.

In the future I would like to talk to the unicode_plot maintainers to see if there's maybe a way to extract core functionality to a Ruby-only gem that does not depend on enumerable-statistics. For example we could name it `unicode_plot_core` and isolate it to cases that don't depend on enumerable-statistics (which is a native C extension).

The gem enumerable-statistics is really only used for quickly generating values like mean, median, mode and calculating histogram values. The core logic of plotting and rendering is loosely coupled, however the integration is via monkey-patches so there is not a easily definable seam.